+++ src/renderer/functors/itextureretrievalfunctor.h
@@ -16,7 +16,7 @@ namespace Renderer
 
         // Get a pointer to a texture by ID.
         // If the ID is 0, return an error texture.
-        virtual OpenGLTexturePointer operator ()(quint64 textureId) const = 0;
+        virtual OpenGLTexturePointer operator ()(quint32 textureId) const = 0;
     };
 }
